Refactored engine throttling
